+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Removed
 - Awaitable Coroutines package dependency
+- Function SceneManager.QuitGame
 
 ## [4.3.0] - 2025-05-14
 ### Changed